# docs/api/openapi-stage1.yaml
openapi: 3.0.3
info:
  title: KPK Platform API (Stage 1)
  version: 0.1.0
servers:
  - url: https://example.org/api
paths:
...
  /atoms:
    get:
      summary: Список атомов знаний
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/KnowledgeAtom' }
    post:
      summary: Создать атом знания
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KnowledgeAtomCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KnowledgeAtom' }
  /atoms/{id}:
    get:
      summary: Получить атом
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KnowledgeAtom' }
    patch:
      summary: Обновить атом
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KnowledgeAtomUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KnowledgeAtom' }
    delete:
      summary: Удалить атом
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
  /contexts:
    get:
      summary: Список профилей контекста
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ContextProfile' }
    post:
      summary: Создать профиль контекста
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContextProfileCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContextProfile' }
  /contexts/{id}:
    get:
      summary: Получить профиль контекста
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContextProfile' }
    patch:
      summary: Обновить профиль контекста
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContextProfileUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContextProfile' }
    delete:
      summary: Удалить профиль контекста
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
...
  /questions/import_xlsx:
    post:
      summary: Импорт MCQ из XLSX (4 варианта, 1 верный)
      description: |
        Ожидается таблица с колонками ``prompt``, ``option1``..``option4``, ``correct`` (номер варианта или текст),
        ``norm_unit_id`` и опционально ``chunk_id``, ``valid_from``, ``valid_to``.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MCQImportResult' }
...
components:
  schemas:
    Citation:
      type: object
      properties:
        norm_unit_id:
          type: string
          nullable: true
        chunk_id:
          type: string
          nullable: true
    KnowledgeAtomBase:
      type: object
      required: [title, statement, type, bloom, granularity]
      properties:
        title: { type: string }
        statement: { type: string }
        type:
          type: string
          enum: [declarative, procedural, conceptual, conditional, metacognitive, normative]
        bloom:
          type: string
          enum: [remember, understand, apply, analyze, evaluate, create]
        granularity:
          type: string
          enum: [micro, meso, macro]
        status:
          type: string
          enum: [draft, reviewed, approved, deprecated]
          default: draft
        version: { type: string, nullable: true }
        prerequisites:
          type: array
          items: { type: string }
        misconceptions:
          type: array
          items: { type: string }
        citations:
          type: array
          items: { $ref: '#/components/schemas/Citation' }
    KnowledgeAtomCreate:
      allOf:
        - type: object
          required: [id]
          properties:
            id: { type: string }
        - $ref: '#/components/schemas/KnowledgeAtomBase'
    KnowledgeAtomUpdate:
      type: object
      properties:
        title: { type: string }
        statement: { type: string }
        type: { type: string }
        bloom: { type: string }
        granularity: { type: string }
        status: { type: string }
        version: { type: string }
        prerequisites:
          type: array
          items: { type: string }
        misconceptions:
          type: array
          items: { type: string }
        citations:
          type: array
          items: { $ref: '#/components/schemas/Citation' }
    KnowledgeAtom:
      allOf:
        - type: object
          properties:
            id: { type: string }
            created_at: { type: string, format: date-time, nullable: true }
            updated_at: { type: string, format: date-time, nullable: true }
        - $ref: '#/components/schemas/KnowledgeAtomBase'
    ContextProfileBase:
      type: object
      required: [title, type, bloom, granularity]
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        type: { type: string }
        bloom: { type: string }
        granularity: { type: string }
        status:
          type: string
          enum: [draft, reviewed, approved, deprecated]
          default: draft
        version: { type: string, nullable: true }
        prerequisites:
          type: array
          items: { type: string }
        misconceptions:
          type: array
          items: { type: string }
        citations:
          type: array
          items: { $ref: '#/components/schemas/Citation' }
        atom_ids:
          type: array
          items: { type: string }
    ContextProfileCreate:
      allOf:
        - type: object
          required: [id]
          properties:
            id: { type: string }
        - $ref: '#/components/schemas/ContextProfileBase'
    ContextProfileUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        type: { type: string }
        bloom: { type: string }
        granularity: { type: string }
        status: { type: string }
        version: { type: string }
        prerequisites:
          type: array
          items: { type: string }
        misconceptions:
          type: array
          items: { type: string }
        citations:
          type: array
          items: { $ref: '#/components/schemas/Citation' }
        atom_ids:
          type: array
          items: { type: string }
    ContextProfile:
      allOf:
        - type: object
          properties:
            id: { type: string }
            created_at: { type: string, format: date-time, nullable: true }
            updated_at: { type: string, format: date-time, nullable: true }
        - $ref: '#/components/schemas/ContextProfileBase'
...
    MCQImportResult:
      type: object
      properties:
        imported: { type: integer }
        rejected_rows:
          type: array
          items: { type: integer }